requirements.txt
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
django-filter==22.1
